move a composable into composables.ts
