+++ src/lib/utils.ts
@@ -40,35 +40,6 @@ export async function ambilGambarBukuDariISBN(isbn: Buku["no_isbn"]) {
   return "assets/Image_not_available.png"
 }
 
-export interface PeminjamanState {
-  id?: Peminjaman["id"]
-  isBorrowable: boolean
-  isReturnable: boolean
-}
-
-export async function usePeminjamanState(isbn: Buku["no_isbn"]): Promise<PeminjamanState> {
-  const peminjamanQuery = supabase
-    .from("peminjaman")
-    .select("id, no_isbn, tgl_pinjam, tgl_kembali, peminjaman_state(id, name)")
-    .eq("no_isbn", isbn)
-    .order("tgl_pinjam", { ascending: false })
-    .limit(1)
-    .maybeSingle()
-
-  const { data, error } = await peminjamanQuery
-  if (error) {
-    throw error
-  }
-
-  if (!data || !data.peminjaman_state) return { isBorrowable: true, isReturnable: false }
-
-  const borrowableConditions = ["borrow cancelled", "return confirmed", "return late"]
-  const isBorrowable = borrowableConditions.includes(data.peminjaman_state.name)
-  const isReturnable = data.peminjaman_state.name === "borrow confirmed"
-
-  return { id: data.id, isBorrowable, isReturnable }
-}
-
 export async function pinjamBukuDariISBN(no_isbn: Buku["no_isbn"], tenggat_waktu: Date) {
   const { error } = await supabase.from("peminjaman").insert({
     no_isbn,
